requirements.txt
google-genai>=0.3.0
python-dotenv>=1.0.0
pyyaml>=6.0
pandas>=2.0.0
